lowering: move lower_expr -> expr.rs
